Rename common_software to census
